backend/requirements.txt
flask
flask-socketio
paho-mqtt
scapy
